Building data is in an asset
This is a large commit as I've neglected to do so in a while.

I've been working on further decoupling the data from the buildables. The building class now holds a serialized object wrapper for the building data. Meaning we can plug in different building data as well as apply the same building data to multiple buildings